xum1541: Added teensy2 board.
The PJRC Teensy 2.0 is a very small, low-cost board with an ATMEGA32U4
controller. It's available from http://www.pjrc.com/teensy/

No additional components are required - all the IEC lines can be
soldered directly to the board.

The hex file was compiled with avr-gcc 4.7.2.

* IEC serial mode has been verified by various people since my original
patch from 2010-03-24 and works fine with 1541 drives.

* IEC parallel mode was verified by me for 1541 drives.

* SRQ nibble support was ported from board-zoomfloppy, and has
  _not_ been verified yet (because I don't have a 1571).